Add list item for pictures grid and image error sad icon
Android layout list_item_pictures_grid.xml is the list item for the PicturesGridActivity that will be set in views for Recycler View. It has an image view for NASA Picture image and text view for name wrapped inside a CardView.
Drawable icon ic_sad is the Image load failure error icon