Add PicturesGridRecyclerAdapter to load data in recycler view
PicturesGridRecyclerAdapter provides a binding for NASA Picture image and name, set to views that are displayed within the recyclerView for the PicturesGridActivity, i.e, Pictures grid screen. Images will be load using Glide will CircularProgressDrawable as placeholder and sad icon for image loading error.
PictureDetails variable names for imageUrl and imageUrlHd changed to url and hdurl respectively to match the json data object.